SAMZA-1048 : upgrade jetty dependency to Jetty 9 from Jetty 8
Jetty 8 is a very old jetty version and the current widely used version is Jetty 9. If a user is using standalone Samza in a Jetty container, and he/she is using Jetty 9, he/she may see some incompatibility issue and it makes a lot of sense for a user to upgrade the Jetty version on his/her side instead of downgrading the Jetty version.

Author: Fred Ji <fji@linkedin.com>

Reviewers: Yi Pan <nickpan47@gmail.com>

Closes #20 from fredji97/master