SAMZA-1123; Create intermediate stream in partitionBy() operator
For partitionBy() operator, Samza generates an intermediate stream with id based on operator name and id, and system based on config. The intermediate streams will be materialized later by different execution environments. For example, if the intermediate stream is a Kafka stream, the topic will be created before the application starts.

Also renamed the config from "job.runner.class" to "app.runner.class".

Author: Xinyu Liu <xiliu@xiliu-ld.linkedin.biz>

Reviewers: Prateek Maheshwari <pmaheshwari@linkedin.com>

Closes #79 from xinyuiscool/SAMZA-1123