SAMZA-1174; Profiling state store performance
This is s commit for [SAMZA-1174](https://issues.apache.org/jira/browse/SAMZA-1174). This commit involves gathering a log of operations (read, write, delete, etc.) happening on the state and publishing them into a kafka topic. It is names as "Access Log" behaving similar to changelog, but gathering log information.

Author: jmehar2 <jmehar2@illinois.edu>
Author: Jayasi Mehar <jayasi05@gmail.com>
Author: s-noghabi <abdolla2@illinois.edu>

Reviewers: Prateek Maheshwari <pmaheshw@linkedin.com>, Jagadish V <jagadish@apache.org>

Closes #132 from s-noghabi/master