Internal cleanup for High-Level API implementation.
Separated the OperatorSpec and MessageStream DAGs so that they're now duals of each other. Users interact with and construct the MessageStream DAG; we create and use the OperatorSpec DAG internally.
Moved common OperatorSpec functionality (getId, getOpCode, getOpName etc.) to the OperatorSpec abstract base class.
New JoinOperatorSpec and PartialJoinOperatorImpls created from JoinOperatorSpec in OperatorGraphImpl.
New InputOperatorSpec and InputOperatorImpl (previously RootOperatorImpl) created when StreamGraph#getInputStream is called.
New OutputOperatorSpec and OutputOperatorImpl for partionBy and sendTo. These are Separate from SinkOperatorSpec for and SinkOperatorImpl for sink. Don't need to create a sinkFn for partitionBy and sendTo anymore.
Updated most unit tests to use the new classes and avoid reflection.