SAMZA-1312: Add Control Messages and Intermediate Stream Serde
In this patch, we add the control message types which includes:
* EndOfStreamMessage
* WatermarkMessage

To support in-band data and control messages, we provide a wrapper serde (IntermediateMessageSerde) to serialize/deserialize data/control messages based on message type byte (first byte in the intermediate stream message). The format of the message is defined in SAMZA-1312. The patch integrates this serde with SerdeManager.

Tested in example jobs deployed locally and works as expected.

Author: Xinyu Liu <xiliu@xiliu-ld.linkedin.biz>

Reviewers: Jagadish V <jagadish@apache.org>

Closes #207 from xinyuiscool/SAMZA-1312