SAMZA-1419; Refactor StreamProcessor tests to remove Thread.sleep()
Refactoring the tests to enable parallel execution. Also, removed unnecessary sleeping during execution. This PR will serve as a prototype to get some feedback on this testing pattern which eliminates state sharing between the tests and opens up the avenue for parallel execution.

This change brings down the test time for these 4 tests to **14 sec** compared to **44 secs**.

Author: Bharath Kumarasubramanian <bkumaras@linkedin.com>

Reviewers: Jagadish <jagadish@apache.org>

Closes #291 from bharathkk/master