SAMZA-1417: Clear and recreate intermediate and metadata streams for batch processing
For each run of a batch application, we need to clear the internal streams from the previous run and recreate new ones. This patch introduces the following:
1) bounded flag in StreamSpec
2) app.mode (BATCH/STREAM) in the application config. An application is a batch app iff all the input streams are bounded.
3) app.runId and use it to generate the internal topics for each run.

run.id generation is not addressed in this pr. There will be another patch to resolve it for both yarn and standalone. For now, this patch only works for yarn.

Author: Xinyu Liu <xiliu@xiliu-ld1.linkedin.biz>

Reviewers: Jake Maes <jmakes@apache.org>

Closes #297 from xinyuiscool/SAMZA-1417