SAMZA-1505: Fix CheckpointTool writing only one ssp per task
Currently when using CheckpointTool to write checkpoints, it only writes a checkpoint of a single ssp per task. By debugging the code, looks like the flatMap() on the checkpoint of Optional tuple(taskname -> Map(ssp -> offset)) merges the results by key taskname. This patch stores the results explicitly in a list and then groupBy() on it, which fixes the problem.

Author: xiliu <xiliu@xiliu-ld1.linkedin.biz>

Reviewers: Jake Maes <jmakes@gmail.com>

Closes #364 from xinyuiscool/SAMZA-1505