SAMZA-1498: Support arbitrary system clock timer in operators
This patch adds the capability to register arbitrary timers for both high-level and low-level api.
For high-level, InitableFunction will pass the TimerRegistry to user through the new OpContext, and user will implement the TimerFunction to get timer notifications.
For low-level api, user can register timer in the TaskContext, and then implements the TimerCallback for specific timer actions.

Author: xiliu <xiliu@linkedin.com>
Author: xinyuiscool <xinyuliu.us@gmail.com>
Author: xinyuiscool <xiliu@linkedin.com>

Reviewers: Pateek M <prateekm@gmail.com>

Closes #419 from xinyuiscool/SAMZA-1498