sql: initial insert support for table destination
Table is another main type of IO abstraction in Samza which supports
both read and write (optional). For the table that do support writes, we
should be able to allow SamzaSQL users to write a query to do that. One
example is to insert into a database. The current code only supports
inserting into a stream. This change adds the initial support for table
insert operation.