Infinite loop when trying to use SamzaSqlApplicationRunner in yarn mode
Right now when we try to use the SamzaSqlApplicationRunner in yarn mode it goes into the infinite loop because the appRunnerConfig that we try to set is being overwritten by the appRunnerConfig passed by the job and SamzaSqlApplicationRunner keeps creating it again and again in an infinite loop.

Fix : Userconfig for app.runner.class should not override the computed one. Added a test case to validate this.

Author: Srinivasulu Punuru <spunuru@linkedin.com>

Reviewers: Xinyu Liu <xinyuliu.us@gmail.com>

Closes #440 from srinipunuru/bug-fix.1