ScheduleAfterDebounceTime should catch all Throwable to avoid losing unhandled Errors
If an Action executed in the scheduler throws an Error (or other Throwable?) besides Exception, it is silently lost since the Action/Runnable wrapper only catches Exception, not Throwable. This made my troubleshooting of an issue very difficult. Made it seem like the code was "hung" when really it had thrown a "NoSuchMethodError" (Error instead of Exception) due to a simple dependency issue on my side.

Catching Throwable instead ensures this is handled and propagated properly.

Author: thunderstumpges <tstumpges@ntent.com>

Reviewers: Prateek Maheshwari <pmaheshwari@apache.org>

Closes #450 from thunderstumpges/scheduler-catch-throwable