SAMZA-1627: Watermark broadcast enhancements
Currently each upstream task needs to broadcast to every single partition of intermediate streams in order to aggregate watermarks in the consumers. A better way to do this is to have only one downstream consumer doing the aggregation, and then broadcast to all the partitions. This is safe as we can prove the broadcast watermark message is after all the upstream tasks finished producing the events whose event time are before the watermark. This reduced the full message count from O(n^2) to O(n).

Author: xinyuiscool <xiliu@linkedin.com>

Reviewers: Boris S <sborya@gmail.com>

Closes #456 from xinyuiscool/SAMZA-1627