SAMZA-1666: Benchmark SystemProducer and SystemConsumer
* Tests to benchmark the performance of the system consumers and producers.
* Config to test the benchmark for the event hub system producer and consumer.

SystemConsumerBench and SystemProducerBench provides base generic implementation to test the benchmark for the system producers and consumers. Any new system that needs benchmark test needs a properties file.

The benchmark test itself is single threaded in the way it consumes and produces events. Scaling the benchmark tests right now involves running multiple processes of these tests in parallel.

Right now we just calculate the event rate, But in future we could create a logging metrics registry to hookup other metrics and log them in console along with event rate while the benchmark tests are being run.

Author: Srinivasulu Punuru <spunuru@linkedin.com>

Reviewers: Jagadish <jagadish@apache.org>, Wei Song<wsong@linkedin.com>

Closes #473 from srinipunuru/benchmark.1