SAMZA-1656: EventHubSystemAdmin does not fetch metadata for valid streams.
Currently successive invocation of EventHubSystemAdmin.getSystemStreamMetadata with the same stream collection returns empty results.

This is an implementation bug, where the streams requested as a part of prior invocations of EventHubSystemAdmin.getSystemStreamMetadata are ignored(due to stale caching).

This bug causes the JobModel generation phase to fail and kills the StreamProcessor.

Author: Shanthoosh Venkataraman <svenkataraman@linkedin.com>

Reviewers: Jagadish <jagadish@apache.org>, Srini P<spunuru@linkedin.com>

Closes #474 from shanthoosh/EventHubSystemAdminBug