SAMZA-1582: removed the yajl-ruby dependency for security reasons
The original ticket suggested updating the yajl-ruby dependency for security reasons.
Many of the ruby dependencies for building the docs site are years old. After upgrading the dependencies, the yajl-ruby dependency just went away.
The dependencies were upgraded to the highest version that was compatible with ruby 2.0.0 (jekyll was the only dependency which has newer versions that are only compatible with ruby versions greater than 2.0.0).

Testing done:
bundle exec jekyll serve --watch --baseurl ""
_docs/local-site-test.sh
sanity checked a couple of links

Author: Cameron Lee <calee@linkedin.com>

Reviewers: Jagadish <jagadish@apache.org>

Closes #486 from cameronlee314/yajlruby