SAMZA-1790: LocalContainerRunner should not extend AbstractApplicationRunner.
LocalContainerRunner is the launcher for the process running SamzaContainer in YARN. It extends the AbstractApplicationRunner since the container was using ApplicationRunner#getStreamSpec to create StreamSpecs from config to create the High Level API DAG. It doesn't implement any of the other APIs from the ApplicationRunner.

With SAMZA-1659 and SAMZA-1745, SamzaContainer no longer needs access to StreamSpec to create and execute the High Level API DAG. We can now clean up the LocalContainerRunner implementation so that it doesn't need to implement the ApplicationRunner interface.

Author: Prateek Maheshwari <pmaheshwari@apache.org>

Reviewers: Jagadish Venkatraman <vjagadis1989@gmail.com>, Yi Pan <nickpan47@gmail.com>

Closes #586 from prateekm/lcr-cleanup