SAMZA-1952: StreamPartitionCountMonitor for standalone.
This patch adds the capability to detect the partition change of the input streams of a stateless standalone jobs and trigger a re-balancing phase(which will essentially account for new partitions from input stream and distribute it to the live processors of the group).

Existing partition count detection of input streams is broken in yarn for stateful jobs. This will be addressed for both yarn and standalone as a part of #622

Author: Shanthoosh Venkataraman <spvenkat@usc.edu>

Reviewers: Boris Shkolnik <boryas@apache.org>

Closes #726 from shanthoosh/stream_partition_count_monitor_for_standalone