SAMZA-1997: Samza-sql diagnostics - instrument project operator
When the user uses Samza-SQL, they use high level declarative language (SQL) for ease and speed of implementation of their Samza Job. Therefore, monitoring the job should provide metrics at this high/logical level. This is the goal of the Samza-SQL diagnostics project. In this first baby-step, we start with instrumenting the Project operator to provide run-time metrics.

Author: Shenoda Guirguis <sguirgui@sguirgui-ld2.linkedin.biz>

Reviewers: Srinivasulu Punuru <spunuru@linkedin.com>, Aditya Toomula <atoomula@linkedin.com>

Closes #806 from shenodaguirguis/samza-sql-diagnostics