SAMZA-1638: Recreate SystemProducer on KafkaCheckpointManager.writeCheckpoint failures.
Retry loop in the existing `KafkaCheckpointManager` implementation retries using the same `SystemProducer` instance on exception and does not recreate it.

When some irrecoverable exceptions occur within the `SystemProducer`, all the subsequent produce message invocations on the `SystemProducer` instance will fail. This had made the entire retry loop on `KafkaCheckpointManager` pointless.

This patch consists of the following changes:
1. This patch addresses the above problem by recreating the `SystemProducer` instance on failure and adds a unit test to verify the functionality.
2. Minor code cleanup in classes: `TestKafkaCheckpointManager` and `KafkaCheckpointManager`.

Author: Shanthoosh Venkataraman <spvenkat@usc.edu>
Author: Shanthoosh Venkataraman <svenkata@linkedin.com>

Reviewers: Dong Lin <lindong28@gmail.com>

Closes #792 from shanthoosh/kafka_checkpoint_manager_fix